requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2
pytz
PyJWT
python-decouple
